% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_file_metadata}
\alias{get_file_metadata}
\title{Get metadata from file paths}
\usage{
get_file_metadata(x)
}
\arguments{
\item{x}{a character. Path to files.}
}
\value{
a data frame.
}
\description{
Given a vector of file paths, return a data frame with the metadata in their
file names.
}
