% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_data}
\alias{process_data}
\title{Process a file}
\usage{
process_data(x, adate, file_out, var_names, fire_codes = 7:9)
}
\arguments{
\item{x}{a character(1). Path to a NetCDF.}

\item{adate}{a date(1). Date of the given file.}

\item{file_out}{a character(1). Path to the resulting file.}

\item{var_names}{a character. Name of the variables to export in the given
file.}

\item{fire_codes}{an integer. Codes in the data that represent fire. For
example 7, 8, and 9 represent fire in VNP14 data.}
}
\value{
a character(1). The save path given for the output file.
}
\description{
Given a file with fire data, export its contents to a text file.
}
